# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Nome da API",
      "version": "1.0.0",
      "description": "Descrição da API"
    },
    "servers": [
      {
        "url": "http://localhost:8080",
        "description": "Servidor local"
      }
    ],
    "paths": {
      "/token/{userId}": {
        "get": {
          "summary": "Gerar token de autenticação para um usuário",
          "tags": ["Autenticação"],
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "description": "ID do usuário",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Token de autenticação gerado com sucesso",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Usuário não encontrado"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      },
      "/cadastrarPessoa": {
        "post": {
          "summary": "Cadastrar uma nova pessoa",
          "tags": ["Pessoas"],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInput"
                }
              }
            }
          },
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "responses": {
            "201": {
              "description": "Pessoa cadastrada com sucesso"
            },
            "400": {
              "description": "Tipo de usuário inválido"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      },
      "/editaPessoa/{id}": {
        "put": {
          "summary": "Editar uma pessoa existente",
          "tags": ["Pessoas"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID da pessoa",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Pessoa editada com sucesso"
            },
            "400": {
              "description": "Tipo de usuário inválido"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      },
      "/deletarPessoa/{id}": {
        "delete": {
          "summary": "Deletar uma pessoa",
          "tags": ["Pessoas"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID da pessoa",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Usuário deletado com sucesso"
            },
            "400": {
              "description": "Tipo de usuário inválido"
            },
            "404": {
              "description": "Usuário não encontrado"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      },
      "/cadastrarProjeto": {
        "post": {
          "summary": "Cadastrar um novo projeto",
          "tags": ["Projetos"],
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectInput"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Projeto cadastrado com sucesso"
            },
            "400": {
              "description": "Erro de validação do usuário ou tipo inválido"
            },
            "404": {
              "description": "Usuário responsável não encontrado"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      },
      "/editarProjeto/{id}": {
        "put": {
          "summary": "Editar um projeto existente",
          "tags": ["Projetos"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID do projeto",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Projeto editado com sucesso"
            },
            "400": {
              "description": "Erro de validação do usuário ou usuário não é responsável"
            },
            "403": {
              "description": "Usuário não é responsável pelo projeto"
            },
            "404": {
              "description": "Projeto não encontrado"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      },
      "/deletarProjeto/{id}": {
        "delete": {
          "summary": "Deletar um projeto",
          "tags": ["Projetos"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID do projeto",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Projeto deletado com sucesso"
            },
            "404": {
              "description": "Projeto não encontrado"
            },
            "403": {
              "description": "Não é possível excluir projetos com candidaturas"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      },
      "/projetos": {
        "get": {
          "summary": "Obter lista de projetos",
          "tags": ["Projetos"],
          "responses": {
            "200": {
              "description": "Lista de projetos obtida com sucesso"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      },
      "/candidatar": {
        "post": {
          "summary": "Candidatar-se a um projeto",
          "tags": ["Candidaturas"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidaturaInput"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Candidatura realizada com sucesso"
            },
            "400": {
              "description": "Usuário já se candidatou a esse projeto"
            },
            "404": {
              "description": "Projeto não encontrado"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      },
      "/selecionarCandidato": {
        "post": {
          "summary": "Selecionar um candidato",
          "tags": ["Candidaturas"],
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelecionarCandidatoInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Candidato selecionado com sucesso"
            },
            "500": {
              "description": "Ocorreu um erro ao selecionar o candidato"
            }
          }
        }
      },
      "/candidatos": {
        "get": {
          "summary": "Obter lista de candidatos",
          "tags": ["Candidaturas"],
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Lista de candidatos obtida com sucesso",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "candidatos": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/UserInput"
                        }
                      }
                    }
                  }
                }
              }
            },
            "403": {
              "description": "Usuário sem permissão para acessar esta rota"
            },
            "404": {
              "description": "Usuário não encontrado"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      },
      "/candidatosInteressados": {
        "get": {
          "summary": "Obter lista de candidatos interessados",
          "tags": ["Candidaturas"],
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Lista de candidatos obtida com sucesso"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      },
      "/candidatosSelecionados/{idProjeto}": {
        "get": {
          "summary": "Obter lista de candidatos selecionados",
          "tags": ["Candidaturas"],
          "parameters": [
            {
              "name": "idProjeto",
              "in": "path",
              "description": "ID do projeto",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Lista de candidatos selecionados obtida com sucesso"
            },
            "500": {
              "description": "Ocorreu um erro ao listar os candidatos selecionados"
            }
          }
        }
      }
    },
    "components": {
        "schemas": {
          "UserInput": {
            "type": "object",
            "properties": {
              "NameUser": {
                "type": "string"
              },
              "Age": {
                "type": "integer",
                "format": "int32"
              },
              "CPF": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "type": {
                "type": "integer",
                "format": "int32"
              }
            },
            "required": [
              "NameUser",
              "Age",
              "CPF",
              "email",
              "type"
            ]
          },
          "ProjectInput": {
            "type": "object",
            "properties": {
              "ProjectName": {
                "type": "string"
              },
              "ProjectDescription": {
                "type": "string"
              },
              "YearStart": {
                "type": "integer",
                "format": "int32"
              },
              "YearEnd": {
                "type": "integer",
                "format": "int32"
              },
              "IdUserResponsible": {
                "type": "integer",
                "format": "int32"
              }
            },
            "required": [
              "ProjectName",
              "ProjectDescription",
              "YearStart",
              "YearEnd",
              "IdUserResponsible"
            ]
          },
          "CandidaturaInput": {
            "type": "object",
            "properties": {
              "idProjeto": {
                "type": "integer",
                "format": "int32"
              },
              "idCandidato":{
                "type": "integer",
                "format": "int32"
              }
            },
            "required": [
              "idProjeto"
            ]
          },
          "SelecionarCandidatoInput": {
            "type": "object",
            "properties": {
              "idCandidato": {
                "type": "integer",
                "format": "int32"
              },
              "idProjeto": {
                "type": "integer",
                "format": "int32"
              }
            },
            "required": [
              "idCandidato",
              "idProjeto"
            ]
          },
          "ApplicationInput": {
            "type": "object",
            "properties": {
              "IdUser": {
                "type": "integer",
                "format": "int32"
              },
              "IdProject": {
                "type": "integer",
                "format": "int32"
              },
              "Stats": {
                "type": "string"
              }
            },
            "required": [
              "IdUser",
              "IdProject",
              "Stats"
            ]
          }
        },
        "securitySchemes": {
          "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "bearerFormat": "JWT"
          }
        }
      }
    }
    